RDB Shredder: change the package to storage from enrich (closes #3036)
