--- 4-storage/relational-database-shredder/src/test/scala/com.snowplowanalytics.snowplow.enrich.hadoop/jobs/bad/SchemaValidationFailed1Spec.scala
@@ -1,86 +1,0 @@
-/*
- * Copyright (c) 2014-2015 Snowplow Analytics Ltd. All rights reserved.
- *
- * This program is licensed to you under the Apache License Version 2.0,
- * and you may not use this file except in compliance with the Apache License Version 2.0.
- * You may obtain a copy of the Apache License Version 2.0 at http://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the Apache License Version 2.0 is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
- */
-package com.snowplowanalytics.snowplow.enrich
-package hadoop
-package jobs
-package bad
-
-// Scalding
-import com.twitter.scalding._
-
-// Cascading
-import cascading.tuple.TupleEntry
-import cascading.tap.SinkMode
-
-// This project
-import JobSpecHelpers._
-
-// Specs2
-import org.specs2.mutable.Specification
-
-/**
- * Holds the input data for the test,
- * plus a lambda to create the expected
- * output.
- */
-object SchemaValidationFailed1Spec {
-
-  val lines = Lines(
-    """snowplowweb	web	2014-06-01 14:04:11.639	2014-05-29 18:16:35.000	2014-05-29 18:16:35.967	unstruct	2b1b25a4-c0df-4859-8201-cf21492ad61b	114221	clojure	js-2.0.0-M2	clj-0.6.0-tom-0.0.4	hadoop-0.5.0-common-0.4.0		68.42.204.218	1242058182	58df65c46e1ac937	11	437ad25b-2006-455e-b5d8-d664b74df8f3	US	MI	Holland	49423	42.742294	-86.0661						http://snowplowanalytics.com/blog/		https://www.google.com/	http	snowplowanalytics.com	80	/blog/			https	www.google.com	80	/			search	Google													{"schema":"iglu:com.snowplowanalytics.snowplow/unstruct_event/jsonschema/1-0-0","data":{"schema":"iglu:com.snowplowanalytics.snowplow/link_click/jsonschema/1-0-0","data":{"elementClasses":["next"]}}}																			Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/35.0.1916.114 Safari/537.36	Chrome	Chrome		Browser	WEBKIT	en-US	1	1	1	0	1	0	0	0	1	1	24	1241	806	Mac OS	Mac OS	Apple Inc.	America/New_York	Computer	0	1440	900	UTF-8	1226	3874"""
-    )
-
-  val expected = (line: String) =>
-    """|{"line":"%s","errors":[
-          |{"level":"error","schema":{"loadingURI":"#","pointer":""},"instance":{"pointer":""},"domain":"validation","keyword":"required","message":"object has missing required properties ([\"targetUrl\"])","required":["targetUrl"],"missing":["targetUrl"]}
-        |]}""".stripMargin.format(line.replaceAll("\"", "\\\\\"")).replaceAll("[\n\r]","").replaceAll("[\t]","\\\\t")
-}
-
-/**
- * Integration test for the EtlJob:
- *
- * Input data is _not_ in the expected
- * Snowplow enriched event format.
- */
-class SchemaValidationFailed1Spec extends Specification {
-
-  import Dsl._
-
-  "A job which processes input lines with invalid Snowplow enriched events" should {
-    ShredJobSpec.
-      source(MultipleTextLineFiles("inputFolder"), SchemaValidationFailed1Spec.lines).
-      sink[String](MultipleTextLineFiles("outputFolder/atomic-events")){ output =>
-        "not write any atomic-events" in {
-          output must beEmpty
-        }
-      }.
-      sink[String](PartitionedTsv("outputFolder", ShredJob.ShreddedPartition, false, ('json), SinkMode.REPLACE)){ output =>
-        "not write any JSONs" in {
-          output must beEmpty
-        }
-      }.
-      sink[TupleEntry](Tsv("exceptionsFolder")){ trap =>
-        "not trap any exceptions" in {
-          trap must beEmpty
-        }
-      }.
-      sink[String](Tsv("badFolder")){ json =>
-        "write a bad row JSON with input line and error message for each input line" in {
-          for (i <- json.indices) {
-            removeTstamp(json(i)) must_== SchemaValidationFailed1Spec.expected(SchemaValidationFailed1Spec.lines(i)._2)
-          }
-        }
-      }.
-      run.
-      finish
-  }
-}
